Correct connection error managing

+++ src/main/java/remote/ClientController.java
@@ -51,27 +51,48 @@ public class ClientController {
 	private Cipher encryptCipher;
 	private PublicKey publicKey;
 	private ByteString pk = null;
-	
 	private ScravaBlockingStub blockingStub;
+	private String ip;
+	private int port;
+	
 	private ClientController(String ip, int port) {
+		this.ip = ip;
+		this.port = port;
 		try {
 			encryptCipher = Cipher.getInstance("RSA/ECB/PKCS1Padding");
-			ManagedChannel channel = ManagedChannelBuilder.forAddress(ip, port).usePlaintext().build();
-			blockingStub = ScravaGrpc.newBlockingStub(channel);
-			ClientData cacheUser = AppCache.getInstance().user;
-			CipherUpdate cu;
+			attemptConnection();
+		} catch (NoSuchAlgorithmException | NoSuchPaddingException e) { e.printStackTrace(); }
+	}
+	
+	public boolean attemptConnection() {
+		ManagedChannel channel = ManagedChannelBuilder.forAddress(ip, port).usePlaintext().build();
+		blockingStub = ScravaGrpc.newBlockingStub(channel);
+		ClientData cacheUser = AppCache.getInstance().user;
+		CipherUpdate cu;
+		try {
+			cu = blockingStub.startConnection(TokenMessage.newBuilder().setToken(cacheUser.getToken()).setUid(cacheUser.getId()).build());
+		} catch(StatusRuntimeException | NullPointerException e) {
+			AppCache.getInstance().user = null;
+			AppCache.save();
+			UserPanel.INSTANCE.regenerate();
 			try {
-				cu = blockingStub.startConnection(TokenMessage.newBuilder().setToken(cacheUser.getToken()).setUid(cacheUser.getId()).build());
-			} catch(StatusRuntimeException | NullPointerException e) {
-				AppCache.getInstance().user = null;
-				AppCache.save();
-				UserPanel.INSTANCE.regenerate();
 				cu = blockingStub.startConnection(TokenMessage.newBuilder().build());
+			} catch(StatusRuntimeException e1) {
+				System.out.println("No connection to server");
+				channel.shutdown();
+				blockingStub = null;
+				return false;
 			}
+		}
+		try {
 			publicKey = KeyFactory.getInstance("RSA").generatePublic(new X509EncodedKeySpec(cu.getPublicKey().toByteArray()));
-			pk = cu.getPublicKey();
 			encryptCipher.init(Cipher.ENCRYPT_MODE, publicKey);
-		} catch (InvalidKeySpecException | NoSuchAlgorithmException | NoSuchPaddingException | InvalidKeyException e) { e.printStackTrace(); }
+		} catch (InvalidKeySpecException | NoSuchAlgorithmException | InvalidKeyException e) {
+			e.printStackTrace();
+			return false;
+		}
+		pk = cu.getPublicKey();
+		return true;
 	}
 	
 	private void refreshCypher() {
@@ -93,6 +114,8 @@ public class ClientController {
 	}
 	
 	public ClientData login(String name,  String pw) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		try {
 			AppCache.getInstance().user = blockingStub.login(
 					ClientLogin.newBuilder()
@@ -113,6 +136,8 @@ public class ClientController {
 	}
 	
 	public ClientData register(String name, String pw) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		try {
 			AppCache.getInstance().user = blockingStub.register(
 					ClientRegister.newBuilder()
@@ -133,51 +158,66 @@ public class ClientController {
 	}	
 	
 	public Iterator<ObjectDescriptor> getUserProjects(int offset) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		return blockingStub.getProjectList(Query.newBuilder().setQuery("author = " + AppCache.getInstance().user.getId()).setOffset(0).build());
 	}
 	
 	public void logoff() {
+		if(blockingStub == null && !attemptConnection())
+			return;
 		AppCache ac = AppCache.getInstance();
 		blockingStub.deleteToken(TokenMessage.newBuilder().setToken(ac.user.getToken()).setUid(ac.user.getId()).build());
 		ac.user = null;
 		AppCache.save();
 	}
 	
-	public void saveProject(Project p) {
+	public boolean saveProject(Project p) {
+		if(blockingStub == null && !attemptConnection())
+			return false;
 		try (ByteArrayOutputStream baos = new ByteArrayOutputStream();
 				ObjectOutputStream oos = new ObjectOutputStream( baos )){
-		ClientData account = AppCache.getInstance().user;        
-        oos.writeObject(p);
-        oos.close();
-		ObjectDescriptor od = saveProject(
-				AuthoredObject.newBuilder()
-							  .setToken(account.getToken())
-							  .setUid(account.getId())
-							  .setObj(SerializedObject.newBuilder()
-									  				  .setId(p.id)
-									  				  .setName(p.name)
-									  				  .setObj(Base64.getEncoder().encodeToString(baos.toByteArray()))
-									  ).build());
-		p.id = od.getId();
+			ClientData account = AppCache.getInstance().user;        
+	        oos.writeObject(p);
+	        oos.close();
+			ObjectDescriptor od = saveProject(
+					AuthoredObject.newBuilder()
+								  .setToken(account.getToken())
+								  .setUid(account.getId())
+								  .setObj(SerializedObject.newBuilder()
+										  				  .setId(p.id)
+										  				  .setName(p.name)
+										  				  .setObj(Base64.getEncoder().encodeToString(baos.toByteArray()))
+										  ).build());
+			p.id = od.getId();
+			return true;
 		} catch(IOException e) {
 			JOptionPane.showMessageDialog(null, "Unable to save project: " + e.getMessage());
+			return false;
 		}
 	}
 	
-	public void deleteProject(ObjectDescriptor p) {
+	public boolean deleteProject(ObjectDescriptor p) {
+		if(blockingStub == null && !attemptConnection())
+			return false;
 		ClientData cd = AppCache.getInstance().user;
 		blockingStub.deleteProject(AuthoredObject.newBuilder()
 												 .setToken(cd.getToken())
 												 .setUid(cd.getId())
 												 .setObj(SerializedObject.newBuilder().setId(p.getId()))
 												 .build());
+		return true;
 	}
 	
 	public ObjectDescriptor saveProject(AuthoredObject so) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		return blockingStub.saveProject(so);
 	}
 	
 	public Project getProject (int id) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		try (ByteArrayInputStream bais = new ByteArrayInputStream(
 												Base64.getDecoder()
 													  .decode(
@@ -199,11 +239,15 @@ public class ClientController {
 	}
 	public static record Tutorial(String title, String content) {}
 	public Tutorial getTutorial(int id) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		SerializedObject so = blockingStub.getTutorial(Query.newBuilder().setQuery(""+id).build());
 		return new Tutorial(so.getName(), so.getObj());
 	}
 	
 	public Iterator<ObjectDescriptor> getTutorialList(Query q) {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		return blockingStub.getTutorialList(q);
 	}
 
@@ -215,6 +259,8 @@ public class ClientController {
 	}
 	
 	public List<IdObject> getUserTutorials() {
+		if(blockingStub == null && !attemptConnection())
+			return null;
 		List<IdObject> res = new ArrayList<>();
 		List<IdObject> temp = new LinkedList<>();
 		int offset = 0;
@@ -232,7 +278,9 @@ public class ClientController {
 		return res;
 	}
 	
-	public void saveTutorial(int id, String name, String content) {
+	public boolean saveTutorial(int id, String name, String content) {
+		if(blockingStub == null && !attemptConnection())
+			return false;
 		blockingStub.saveTutorial(AuthoredObject.newBuilder()
 												.setUid(AppCache.getInstance().user.getId())
 												.setToken(AppCache.getInstance().user.getToken())
@@ -242,6 +290,7 @@ public class ClientController {
 																		.setObj(content)
 																		.build())
 												.build());
+		return true;
 	}
 	
 }
